+++ Axi4/testbench_MultipleMemory/TbAxi4_MultipleMemory.vhd
@@ -65,14 +65,20 @@ architecture TestHarness of TbAxi4_MultipleMemory is
   signal Clk         : std_logic ;
   signal nReset      : std_logic ;
 
-  signal Manager1Rec, Manager2Rec, Subordinate1Rec, Subordinate2Rec : AddressBusRecType (
+--  signal Manager1Rec, Manager2Rec, Subordinate1Rec, Subordinate2Rec : AddressBusRecType (
+--          Address(AXI_ADDR_WIDTH-1 downto 0),
+--          DataToModel(AXI_DATA_WIDTH-1 downto 0),
+--          DataFromModel(AXI_DATA_WIDTH-1 downto 0)
+--        ) ;
+
+  signal ManagerRec, SubordinateRec : AddressBusRecArrayType(1 to 2)(
           Address(AXI_ADDR_WIDTH-1 downto 0),
           DataToModel(AXI_DATA_WIDTH-1 downto 0),
           DataFromModel(AXI_DATA_WIDTH-1 downto 0)
         ) ;
 
 
-  signal AxiBus1, AxiBus2 : Axi4RecType(
+  signal AxiBus :  Axi4RecArrayType(1 to 2)(
     WriteAddress(
       Addr(AXI_ADDR_WIDTH-1 downto 0),
       ID(7 downto 0),
@@ -140,10 +146,10 @@ begin
     nReset      => nReset,
 
     -- AXI Manager Functional Interface
-    AxiBus  => AxiBus1,
+    AxiBus  => AxiBus(1),
     
     -- Testbench Transaction Interface
-    TransRec    => Subordinate1Rec
+    TransRec    => SubordinateRec(1)
   ) ;
 
   Memory_2 : Axi4Memory
@@ -153,10 +159,10 @@ begin
     nReset      => nReset,
 
     -- AXI Manager Functional Interface
-    AxiBus  => AxiBus2,
+    AxiBus  => AxiBus(2),
     
     -- Testbench Transaction Interface
-    TransRec    => Subordinate2Rec
+    TransRec    => SubordinateRec(2)
   ) ;
 
   Manager_1 : Axi4Manager
@@ -166,10 +172,10 @@ begin
     nReset      => nReset,
 
     -- AXI Manager Functional Interface
-    AxiBus      => AxiBus1,
+    AxiBus      => AxiBus(1),
 
     -- Testbench Transaction Interface
-    TransRec    => Manager1Rec
+    TransRec    => ManagerRec(1)
   ) ;
 
   Manager_2 : Axi4Manager
@@ -179,10 +185,10 @@ begin
     nReset      => nReset,
 
     -- AXI Manager Functional Interface
-    AxiBus      => AxiBus2,
+    AxiBus      => AxiBus(2),
 
     -- Testbench Transaction Interface
-    TransRec    => Manager2Rec
+    TransRec    => ManagerRec(2)
   ) ;
 
 
@@ -192,11 +198,11 @@ begin
     nReset        => nReset,
 
     -- Transaction Interfaces
-    Manager1Rec      => Manager1Rec,
-    Subordinate1Rec  => Subordinate1Rec,
+    Manager1Rec      => ManagerRec(1),
+    Subordinate1Rec  => SubordinateRec(1),
 
-    Manager2Rec      => Manager2Rec,
-    Subordinate2Rec  => Subordinate2Rec
+    Manager2Rec      => ManagerRec(2),
+    Subordinate2Rec  => SubordinateRec(2)
   ) ;
 
 end architecture TestHarness ;