+++ Axi4/testbench_MultipleMemory/TbAxi4_MultipleMemory_Generate.vhd
@@ -65,14 +65,20 @@ architecture TestHarness of TbAxi4_MultipleMemory is
   signal Clk         : std_logic ;
   signal nReset      : std_logic ;
 
-  type AddressBusArrayRecType is array(1 to 2) of AddressBusRecType (
+--  type AddressBusArrayRecType is array(1 to 2) of AddressBusRecType (
+--          Address(AXI_ADDR_WIDTH-1 downto 0),
+--          DataToModel(AXI_DATA_WIDTH-1 downto 0),
+--          DataFromModel(AXI_DATA_WIDTH-1 downto 0)
+--        ) ;
+--  signal ManagerRec, SubordinateRec : AddressBusArrayRecType ; 
+  signal ManagerRec, SubordinateRec : AddressBusRecArrayType(1 to 2)(
           Address(AXI_ADDR_WIDTH-1 downto 0),
           DataToModel(AXI_DATA_WIDTH-1 downto 0),
           DataFromModel(AXI_DATA_WIDTH-1 downto 0)
         ) ;
-  signal ManagerRec, SubordinateRec : AddressBusArrayRecType ; 
 
-  type Axi4ArrayRecType is array(1 to 2) of Axi4RecType(
+
+  signal AxiBus :  Axi4RecArrayType(1 to 2)(
     WriteAddress(
       Addr(AXI_ADDR_WIDTH-1 downto 0),
       ID(7 downto 0),
@@ -99,9 +105,6 @@ architecture TestHarness of TbAxi4_MultipleMemory is
       User(7 downto 0)
     )
   ) ;
-  
-  signal AxiBus : Axi4ArrayRecType ;
-
 
   component TestCtrl is
     port (