+++ common/src/Axi4InterfacePkg.vhd
@@ -21,6 +21,7 @@
 --
 --  Revision History:
 --    Date      Version    Description
+--    09/2024   2024.09    Added Axi4RecArrayType.   Axi4RecType is the record type and no longer an alias
 --    03/2022   2022.03    Factored out of Axi4InterfaceCommonPkg
 --    01/2020   2020.01    Updated license notice
 --    09/2017   2017       Initial revision
@@ -28,7 +29,7 @@
 --
 --  This file is part of OSVVM.
 --  
---  Copyright (c) 2017 - 2022 by SynthWorks Design Inc.  
+--  Copyright (c) 2017 - 2024 by SynthWorks Design Inc.  
 --  
 --  Licensed under the Apache License, Version 2.0 (the "License");
 --  you may not use this file except in compliance with the License.
@@ -104,7 +105,7 @@ package Axi4InterfacePkg is
     InitVal   : std_logic := 'Z'
   ) return Axi4WriteDataRecType ;
 
--- Add VHDL-2018 modes here
+-- Add VHDL-2019 mode views here
 
   
   -- AXI Write Response Channel
@@ -123,7 +124,7 @@ package Axi4InterfacePkg is
     InitVal   : std_logic := 'Z'
   ) return Axi4WriteResponseRecType ;
   
--- Add VHDL-2018 modes here
+-- Add VHDL-2019 mode views here
 
   
   -- AXI Read Address Channel
@@ -155,7 +156,7 @@ package Axi4InterfacePkg is
     InitVal   : std_logic := 'Z'
   ) return Axi4ReadAddressRecType ;
 
--- Add VHDL-2018 modes here
+-- Add VHDL-2019 mode views here
 
   -- AXI Read Data Channel
   type Axi4ReadDataRecType is record
@@ -176,15 +177,20 @@ package Axi4InterfacePkg is
   ) return Axi4ReadDataRecType ;
 
   
--- Add VHDL-2018 modes here
+-- Add VHDL-2019 mode views here
   
-  type Axi4BaseRecType is record
+  type Axi4RecType is record
     WriteAddress   : Axi4WriteAddressRecType ; 
     WriteData      : Axi4WriteDataRecType ; 
     WriteResponse  : Axi4WriteResponseRecType ; 
     ReadAddress    : Axi4ReadAddressRecType ; 
     ReadData       : Axi4ReadDataRecType ; 
-  end record Axi4BaseRecType ; 
+  end record Axi4RecType ; 
+  
+  type Axi4RecArrayType  is array (integer range <>) of Axi4RecType ;
+
+  alias Axi4BaseRecType is Axi4RecType ;  -- mainly since the following subtypes are not well enough supported yet
+  
   
   -- Axi4RecType with sized elements
   -- Get From Above
@@ -208,7 +214,6 @@ package Axi4InterfacePkg is
 --   ReadData     ( Data(open), ID(7 downto 0), User(7 downto 0) )
 -- ) ;
  
-  alias Axi4RecType is Axi4BaseRecType ; 
   
   function InitAxi4Rec (
     AxiBusRec : in Axi4RecType ;